Add native Wii audio driver Added some GBA optimizations (16bpp) Added 16BPP support for PCFX. Added Save State support (10 states) Fixed WonderSwan SRAM memory issue Updated to use standard USB support (need to test this more)

+++ trunk/mednafen/src/gba/Gfx.h
@@ -60,7 +60,7 @@ extern MDFN_ALIGN(16) uint32 line2[512];
 extern MDFN_ALIGN(16) uint32 line3[512];
 extern MDFN_ALIGN(16) uint32 lineOBJ[512];
 extern MDFN_ALIGN(16) uint32 lineOBJWin[512];
-extern MDFN_ALIGN(16) uint32 lineMix[512];
+extern MDFN_ALIGN(16) uint16 lineMix[512];
 extern bool gfxInWin0[512];
 extern bool gfxInWin1[512];
